PandasAI-Streamlit-App-main/requirements.txt
langchain
pandasai
python-dotenv
streamlit




